Changes following Pull request discussion
HardwareObjects.General -> HardwareRepository.ConvertUtils
Changed camelCase function in General to snake_case
Moved States to queue_model_enumerations_v1
