+++ HardwareObjects/queue_model_enumerables_v1.py
@@ -1,8 +1,10 @@
 """
-Enumberables and oheter constants used by the queue model.
+Enumerables and other constants used by the queue model.
 """
 
 from collections import namedtuple
+
+import enum
 
 StrategyComplexity = namedtuple('StrategyComplexity', ['SINGLE','FEW','MANY'])
 STRATEGY_COMPLEXITY = StrategyComplexity('none', 'min', 'full')
@@ -72,3 +74,30 @@ SPACEGROUP_NUMBERS = dict((name.strip(), int(value))
                           for name, value in ORIG_EDNA_SPACEGROUPS.items())
 # For Germanate (test case):
 SPACEGROUP_NUMBERS['Ia-3d'] = '230'
+
+
+@enum.unique
+class States(enum.Enum):
+    """Standard device states, based on TangoShutter states.
+    SardanaMotor.state_map, and DiffractometerState,
+    for general use across HardwareObjects.
+
+    Limited to a common set of states.
+
+    Grammar of tags corrected ('CLOSE->CLOSED, DISABLE->DISABLED) relative
+    to Tango states, so you can echo them to the UI without misunderstanding"""
+
+    CLOSED = 0
+    OPEN = 1  # Also used for workflow 'Expecting input'
+    ON = 2  # Could be used to mean 'Connected'.
+    OFF = 3  # Could be used to mean 'Disconnected'
+    INSERT = 4
+    EXTRACT = 5
+    MOVING = 6
+    STANDBY = 7  # Could be used to mean 'Ready'
+    FAULT = 8
+    INIT = 9
+    RUNNING = 10
+    ALARM = 11
+    DISABLED = 12
+    UNKNOWN = 13
